Add export to demo.
Rename Insert to InsertAlert and add InsertAlertExport. I think
this will make it easier to not typo Insert v. TABLE.insert v.
Insert2 in the demo.

Not that I've done that or anything.

Also add an export table with a further involved procedure
(InsertAlertExport) that mirrors alerts to export table.